docker: add wrapper for all stages
Each of the docker wrappers (app-docker, cli-docker, etc)
can make use of this single script in order to integrate
the various actions that they would like to perform.

In most cases, `. "$(dirname "$0")/docker app"` should
suffice as the entirety of script. With time, these scripts
might should be deprecated.

In travis, the `.omero/app-docker` invocations should still
work, but `.omero/docker app scripts`, for example, should
be usable for a multi-staged build.

To test what the script would do, use:

```
  ACTION=echo .omero/docker app scripts
```