Be stricter about the definition of "$PROJECT"
Although we missed the boat on using COMPOSE_PROJECT_NAME,
our use of PROJECT is fairly clean. To improve it slightly,
this commit makes the default the value of `get_project_name`
which works similarly to `get_app_name` but tries to match
the logic that docker-compose would use. There doesn't look
to be a simple way to inspect the output from `docker-compose`
to get what the project name _would be_. Since we need the
value before docker-compose startup, this seems like a fine
compromise.