Use no ports by default (breaking change)
Running more than one .omero setup on a single host
ends up requiring changes to either .env or to the
environment to avoid port conflicts. If the ports
are randomized by default, only the case where hard-
coded ports are needed requires changing. This is a
breaking change since the ":" now belongs in the .env
file and *not* in the docker-compose.yml file.