IDR: use downloaded ansible templates & variables to configure infra
In order to not reproduce the configuration necessary for setting up
an IDR clone within omero-testinfra, ansible is used to download the
template files as well as the default variables from various roles &
the variables are downloaded from idr/deployment. The performance of
mapr calls is slow without nginx caching, but the timeout looks near
long enough to allow most calls to succeed.