Split server/test responsibilities apart
The Dockerfile in openmicroscopy/openmicroscopy uses two stages:
build (FROM openjdk) and deploy (FROM omero-server). The latter
does not have the necessary executables (like slice2java) to
properly function as an integration test suite.

Therefore we now have a new "test" container in srv-compose.yml
which stops the build during the "build" stage and therefore
keeps the necessary artifacts for testing.

This required upgrading the compose file version to 3.4